applying interface segregation and other PR reviews
