+++ tracing/src/main/kotlin/br/com/guiabolso/tracing/factory/CompositeTracerEngine.kt
@@ -3,7 +3,7 @@ package br.com.guiabolso.tracing.factory
 import br.com.guiabolso.tracing.engine.TracerEngine
 import java.io.Closeable
 
-class CompositeTracerEngine(
+open class CompositeTracerEngine(
         private var tracers: List<TracerEngine<*>>
 ) : TracerEngine<Map<TracerEngine<*>, Any>> {
 
@@ -21,23 +21,6 @@ class CompositeTracerEngine(
 
     override fun addProperty(key: String, value: Boolean?) {
         tracers.forEach { it.addProperty(key, value) }
-    }
-
-    override fun recordExecutionTime(name: String, elapsedTime: Long, context: MutableMap<String, String>) {
-        tracers.forEach {
-            it.recordExecutionTime(name, elapsedTime, context)
-        }
-    }
-
-    override fun <T> executeAndRecordTime(name: String, block: (MutableMap<String, String>) -> T): T {
-        val start = System.currentTimeMillis()
-        val context = mutableMapOf<String, String>()
-        try {
-            return block(context)
-        } finally {
-            val elapsedTime = System.currentTimeMillis() - start
-            recordExecutionTime(name, elapsedTime, context)
-        }
     }
 
     override fun notifyError(exception: Throwable, expected: Boolean) {
